Fix strange NPE, which was fixed before...

+++ delis-web/src/main/java/dk/erst/delis/web/validate/ValidateController.java
@@ -134,8 +134,10 @@ public class ValidateController {
 					resultList.add(result);
 
 				} finally {
-					deleteFile(infoData.getFile());
-					deleteFile(infoData.getFileSbd());
+					if (infoData != null) {
+						deleteFile(infoData.getFile());
+						deleteFile(infoData.getFileSbd());
+					}
 				}
 			}
 		}
